# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "chainmail API",
    "version": "1.0.0",
    "description": "Read-only HTTP over a local chainmail corpus: search for candidate chains, choose which belong on a page, then build the timeline spec the renderer consumes.\n\nTwo stages on purpose. Search selects; POST /v1/spec renders a chosen set. A spec built straight from a query takes whatever the query returned, and then the only way to remove a thread is to regenerate the page.\n\nThis serves personal mail and has no authentication of any kind. It binds to loopback and refuses a non-loopback address unless explicitly overridden, and spec bodies carry unsanitised sender HTML (issue #14) \u2014 so the loopback bind is what makes the whole surface safe, not a default to be relaxed.\n\nOperator commands \u2014 ingest, embed, eval, dedupe, twins, repair, merge, alias, refresh \u2014 are absent in their mailbox-touching forms; that reach belongs to the CLI and the cron. POST /v1/refresh is the one exception and it stays local: it re-derives a saved page from the corpus and, before redrawing, collapses stored twin copies (one message stored twice) with the same refusal-based sweep the CLI's slurp runs."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "The only supported deployment: loopback on the machine holding the corpus."
    }
  ],
  "paths": {
    "/v1/search": {
      "get": {
        "summary": "Ranked chains for a query \u2014 the selection candidates.",
        "description": "Selection stage. Lexical by default: FTS5 wins decisively on the tokens a mail corpus is dense with (invoice numbers, account ids, Message-IDs), and the structural half of a real question is a filter rather than something a similarity score encodes.\n\nAt least one of q, person or since must be given. With no filter at all an arbitrary slice of the corpus would come back reading like a ranked answer, so that is a 400.",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Free text. Double-quoted runs are kept as phrases; every other term must match. Terms that look like identifiers also go to the trigram index.",
            "schema": {
              "type": "string",
              "description": ""
            },
            "example": "solar install quote"
          },
          {
            "name": "person",
            "in": "query",
            "description": "An address, display name or Slack uid. Matches anyone recorded on the entry, author or recipient \u2014 on a real trail a quarter of the participants appear only in to:/cc:, so \"threads Ada was on\" cannot be an author-only filter.",
            "schema": {
              "type": "string",
              "description": ""
            },
            "example": "ada@example.com"
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only entries on or after this date.",
            "schema": {
              "type": "string",
              "description": "",
              "format": "date"
            },
            "example": "2026-01-31"
          },
          {
            "name": "mode",
            "in": "query",
            "description": "Retrieval. semantic and hybrid need vectors (`corpus embed`) AND a running local embedding daemon; without the daemon they return 503, never an empty result set.",
            "schema": {
              "type": "string",
              "description": "",
              "enum": [
                "lexical",
                "semantic",
                "hybrid"
              ],
              "default": "lexical"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many chains (or entries) to return.",
            "schema": {
              "type": "integer",
              "description": "",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          },
          {
            "name": "entries",
            "in": "query",
            "description": "Return the matching entries instead of the chains they belong to. Changes which array of the response is populated.",
            "schema": {
              "type": "boolean",
              "description": "",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked candidates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "No filter given, or a parameter could not be parsed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "mode=semantic or mode=hybrid, and the embedding daemon is not reachable or has not pulled the model. The message names the fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "The embedding daemon did not answer in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/spec": {
      "post": {
        "summary": "Build a timeline spec from a chosen set of chains.",
        "description": "Render stage, and deliberately separate from search: what belongs on a page is a human decision, and removing a thread afterwards means regenerating everything downstream, so scope is settled once, up front.\n\nThe response validates against schema/timeline.schema.json, the contract the renderer already consumes.\n\nExpensive: spec assembly recovers quoted bodies, detects repeated boilerplate and builds attachment thumbnails, which is seconds for a large selection. There is no job queue \u2014 the response is synchronous. Concurrency is capped, so an over-eager caller gets 429 rather than a server with every core in HTML recovery.\n\nEntry bodies are sender HTML, and it is NOT sanitised (issue #14). That is accepted at today's risk because the server binds to loopback only. It is a hard blocker on binding wider, not a footnote.",
        "operationId": "buildSpec",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A timeline spec, valid against schema/timeline.schema.json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineSpec"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body, an empty chains list, or a selection that resolved to no entries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "One of the named chains is not in the corpus. The message names which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many spec builds already in flight. Retry-After is set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specs/{name}": {
      "get": {
        "summary": "A page that POST /v1/spec saved under a name — the read half of the render route /view/<name>.",
        "description": "The page was written by the same server process family (specs live beside the corpus, in ~/.local/state/chainmail/specs), so a refresh or a reboot lands back on the same page from the same URL. The response is byte-identical to what the build returned.",
        "operationId": "getSavedSpec",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name the build was saved under. Letters, digits, '.', '_', '-'; no slashes and no '..'.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
            },
            "example": "solar-install-quote"
          }
        ],
        "responses": {
          "200": {
            "description": "The saved timeline spec.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineSpec"
                }
              }
            }
          },
          "400": {
            "description": "The name is not a valid spec name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No page was ever saved under that name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entries/{extId}": {
      "get": {
        "summary": "One entry, with its provenance and every sighting of it.",
        "operationId": "getEntry",
        "parameters": [
          {
            "name": "extId",
            "in": "path",
            "description": "Corpus natural key, as search reports it. Percent-encode it: it contains a colon and usually angle brackets.",
            "required": true,
            "schema": {
              "type": "string",
              "description": ""
            },
            "example": "mail:<c0ffee-1@example.com>"
          }
        ],
        "responses": {
          "200": {
            "description": "The entry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusEntry"
                }
              }
            }
          },
          "404": {
            "description": "No entry carries that id. Distinct from a search that matched nothing, which is a 200 with an empty array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SearchResponse": {
        "type": "object",
        "description": "Ranked candidates. Exactly one of chains or entries is present, decided by the `entries` parameter; neither is present when nothing matched, and an empty array is returned rather than omitted in that case.",
        "required": [
          "mode"
        ],
        "additionalProperties": false,
        "properties": {
          "mode": {
            "type": "string",
            "description": "The retrieval mode actually used.",
            "enum": [
              "lexical",
              "semantic",
              "hybrid"
            ]
          },
          "chains": {
            "type": "array",
            "description": "Present when entries=false (the default).",
            "items": {
              "$ref": "#/components/schemas/ChainHit"
            }
          },
          "entries": {
            "type": "array",
            "description": "Present when entries=true.",
            "items": {
              "$ref": "#/components/schemas/EntryHit"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "Every non-2xx response carries this and nothing else.",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in the same words the CLI would use. Written to be shown to a person.",
            "example": "mode \"semantic\": no embedding daemon reachable at http://localhost:11434: start it with `ollama serve`"
          }
        }
      },
...
    }
  }
}